// src/app/interfaces/user.ts
export interface User {
    uid: string;
    firstName: string;
    lastName: string;
    email: string;
    password: string;
    avatar: string;
}
export interface UserAccount {
    uid: string;
    email: string;
    firstName: string;
    lastName: string;
    password: string;
    avatarUrl : string;
  }